# apps/server/build/swagger.yaml
openapi: 3.0.0
components:
  examples: {}
  headers: {}
  parameters: {}
  requestBodies: {}
  responses: {}
  schemas:
    AgentModelOption:
      description: >-
        Curated list of LLM models the user can pick from. The `id` is the

        OpenRouter model slug forwarded to the agent's `/agent/respond`
        endpoint.

        Surface UX assumes 4-6 options.
      properties:
        id:
          type: string
        label:
          type: string
        description:
          type: string
      required:
        - id
        - label
        - description
      type: object
      additionalProperties: false
    UserPreferencesDto:
      properties:
        preferredModel:
          type: string
      required:
        - preferredModel
      type: object
      additionalProperties: false
    PatchUserPreferencesBody:
      properties:
        preferredModel:
          type: string
      type: object
      additionalProperties: false
    ChatListItemDto:
      properties:
        id:
          type: string
        title:
          type: string
        starred:
          type: boolean
        isPublic:
          type: boolean
        updatedAt:
          type: string
      required:
        - id
        - title
        - starred
        - isPublic
        - updatedAt
      type: object
      additionalProperties: false
    CmuMapsDto:
      properties:
        url:
          type: string
          nullable: true
        mode:
          type: string
          nullable: true
        target:
          type: string
          nullable: true
        targetLabel:
          type: string
          nullable: true
        src:
          type: string
          nullable: true
        srcLabel:
          type: string
          nullable: true
        dest:
          type: string
          nullable: true
        destLabel:
          type: string
          nullable: true
      required:
        - url
        - mode
        - target
        - targetLabel
        - src
        - srcLabel
        - dest
        - destLabel
      type: object
      additionalProperties: false
    MessageDto:
      properties:
        id:
          type: string
        role:
          type: string
          enum:
            - user
            - assistant
            - system
        content:
          type: string
        createdAt:
          type: string
        cmuMaps:
          allOf:
            - $ref: "#/components/schemas/CmuMapsDto"
          nullable: true
        confidence:
          type: number
          format: double
          description: >-
            Agent confidence for the just-generated turn. Not persisted; only
            set on

            fresh assistant messages, undefined when re-reading history.
      required:
        - id
        - role
        - content
        - createdAt
      type: object
      additionalProperties: false
    ChatDetailDto:
      properties:
        id:
          type: string
        title:
          type: string
        starred:
          type: boolean
        isPublic:
          type: boolean
        updatedAt:
          type: string
        isOwner:
          type: boolean
      required:
        - id
        - title
        - starred
        - isPublic
        - updatedAt
        - isOwner
      type: object
      additionalProperties: false
    PostMessageResultDto:
      properties:
        userMessage:
          $ref: "#/components/schemas/MessageDto"
        assistantMessage:
          $ref: "#/components/schemas/MessageDto"
      required:
        - userMessage
        - assistantMessage
      type: object
      additionalProperties: false
    PostChatMessageBody:
      properties:
        content:
          type: string
      required:
        - content
      type: object
      additionalProperties: false
    PatchChatBody:
      properties:
        starred:
          type: boolean
        title:
          type: string
        isPublic:
          type: boolean
      type: object
      additionalProperties: false
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://idp.scottylabs.org/realms/scottylabs/.well-known/openid-configuration
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your bearer token below
    oidcOrBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Better Auth session cookie and/or Keycloak access token as Bearer;
        server merges both.
info:
  title: TSOA
  version: 1.0.0
  description: Build swagger-compliant REST APIs using TypeScript and Node
  license:
    name: MIT
  contact:
    name: unknown
paths:
  /me/oidc-admin:
    get:
      operationId: GetOidcAdminStatus
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  isOidcAdmin:
                    type: boolean
                required:
                  - isOidcAdmin
                type: object
      security:
        - oidc: []
      parameters: []
  /me/models:
    get:
      operationId: ListModels
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  models:
                    items:
                      $ref: "#/components/schemas/AgentModelOption"
                    type: array
                required:
                  - models
                type: object
      description: Curated list of LLM models the user can pick from.
      security:
        - oidc: []
      parameters: []
  /me/preferences:
    get:
      operationId: GetPreferences
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferencesDto"
      description: Read the user's preferences (preferred model, etc.).
      security:
        - oidc: []
      parameters: []
    patch:
      operationId: UpdatePreferences
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferencesDto"
      description: Update the user's preferences. Only fields present in the body are
        changed.
      security:
        - oidc: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchUserPreferencesBody"
  /hello:
    get:
      operationId: GetHello
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
      security: []
      parameters: []
  /hello/authenticated:
    get:
      operationId: GetHelloAuthenticated
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
      security:
        - oidc: []
      parameters: []
  /hello/admin:
    get:
      operationId: GetHelloAdmin
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                required:
                  - message
                type: object
      security:
        - oidc: []
      parameters: []
  /chats:
    get:
      operationId: ListChats
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ChatListItemDto"
                type: array
      security:
        - oidc: []
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
    post:
      operationId: CreateChat
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatListItemDto"
      security:
        - oidc: []
      parameters: []
  /chats/{id}/messages:
    get:
      operationId: GetMessages
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/MessageDto"
                type: array
      security:
        - oidc: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
    post:
      operationId: PostMessage
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostMessageResultDto"
      security:
        - oidc: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostChatMessageBody"
  /chats/{id}:
    get:
      operationId: GetChat
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatDetailDto"
      security:
        - oidc: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
    patch:
      operationId: PatchChat
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatListItemDto"
      security:
        - oidc: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchChatBody"
    delete:
      operationId: DeleteChat
      responses:
        "204":
          description: ""
      security:
        - oidc: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
servers:
  - url: /